Allow to link tests with arbitrary libraries.
I implemented this in the same way I normally import C files as Python
modules -- each file starts with a header that contains a number of
`//!`-directives, each specifying a library dir, an include dir,
a library to link with, or an additional source code file. These
directives are reformatted as compiler options and passed to $(CXX).