Remove all mention of backreferences.
So I've been thinking, how the hell did that even work. Turns out,
the answer? It did not. These weren't "true" backreferences --
instead, the NFA matched the groups the best way it could, and
when the backreference failed, it simply did not consider that
maybe there's a different arrangement of group boundaries
that would not make it fail -- from the point of view of the NFA,
all threads that ended up in the same state are equivalent, so it
can choose the one with the highest priority and discard the rest,
group alignment not considered important.