Allow thread_dispatch to work in multiple different modes.
Instead of assuming states are described by functions, it now accepts
pointers to any data. Interpreting this data is done by calling
a single pointer with the NFA and the data as arguments.

No idea how to make a pointer to a method of an instance, though.
So the VM does not work. Yet. The x86-64 backend did not get
any slower, though, so that's good.