vm: in presence of backreferences, use bitmaps selectively.
This makes the code pass all backreference tests.

A reminder: state bitmaps can only be shared between threads which
have the same contents of referenced captured groups. To do that,
each thread has a "version" which is incremented every time
the location of a backreferenced group changes. If we notice
that a thread has a different version than the previous one, we
clear the state bitmap.