Add a markdown-like markup parser as a test.
On my Chromebook, it runs 20-30% faster with re2jit than with re2.
And that is after I replaced `[\p{N}\p{L}]` with `\w` -- the former
makes re2 (but not re2jit!) unbearably slow. Oh, and I had to slightly
restrict the syntax for `inline code` (which originally allowed
````for any number of wrapping backticks````) because re2 does not
support backreferences.