Store small bitmaps on stack.
If the bitmap is sizeof(size_t) bytes or shorter, we can easily
fit it into the NFA struct itself, which allows us to do one less
memory lookup to see if a bit is set, as well as clear the whole
bitmap with a single assignment instead of a `memset`.

This provides marginal (~1-3%) performance increase at the cost
of making implementaton more complex (and the VM does not work, again).